Refactor units.py file location and update import paths
The file units.py has been moved out of the base folder up to src/mdio/schemas directory, subsequently leading to changes in all the import paths that pointed at it. This restructuring simplifies the directory structure and improves the accessibility to units.py file.
